+++ stars-importer-carla/src/main/kotlin/tools/aqua/stars/importer/carla/dataclasses/JsonLaneMarking.kt
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2026 The STARS Project Authors
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package tools.aqua.stars.importer.carla.dataclasses
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+/**
+ * Json object for a lane marking on one side of a [JsonLane].
+ *
+ * @property markingType The [JsonLaneMarkingType] of the marking.
+ * @property color The [JsonLaneMarkingColor] of the marking.
+ * @property width The width of the marking in meters.
+ */
+@Serializable
+data class JsonLaneMarking(
+    @SerialName("marking_type") val markingType: JsonLaneMarkingType,
+    @SerialName("color") val color: JsonLaneMarkingColor,
+    @SerialName("width") val width: Double,
+)